+++ docs/change_log.rst
@@ -1,5 +1,17 @@
 更新记录
 ===========================
+
+2020.3.28 v1.2
+---------------------------
+* #255 #266, 开始使用 github 的 Actions 进行 CI 集成;
+* `#257 <https://github.com/chinapnr/fishbase/issues/257>`_, conf_as_dict() 函数移动从 common 包移动到 system 包;
+* `#259 <https://github.com/chinapnr/fishbase/issues/259>`_, 增加 `fish_object <https://fishbase.readthedocs.io/en/latest/fish_object.html>`_ 函数包，面向对象增强函数;
+* `#260 <https://github.com/chinapnr/fishbase/issues/261>`_, common 包，删除 sorted_objs_by_attr() 函数和 get_group_list_data() 函数;
+* `#261 <https://github.com/chinapnr/fishbase/issues/261>`_, common 包，paging 函数名称修改为 :meth:`get_page_data()`;
+* `#263 <https://github.com/chinapnr/fishbase/issues/263>`_, common 包，删除一些为了向前兼容的函数;
+* `#263 <https://github.com/chinapnr/fishbase/issues/265>`_, 删除 flask swagger 支持;
+
+
 2019.12.5 v1.1.16
 ---------------------------
 * `#249 <https://github.com/chinapnr/fishbase/issues/249>`_, add swagger module;